Add overwrite option to overwrite even if path is duplicated
SUPPOSED USE CASE:

This should be useful to use together with `%{uuid_chunk}` placeholder.
`%{uuid_chunk}` placeholder assures that the path is unique for a chunk,
hence, path duplication error occurrence means that a mistaken retry is
happened. In such case, we can simply remove the old file by
overwriting with the new file.